Add mock user data, refcator routing service
